Release Automation (#15)
* Release Automation

This adds a GitHub Action that packages up the syntax files for consumption in other products. The action is triggered by a git tag matching the version strategy we use and automatically generates a changelog inside the GitHub release based on all pull requests since last. It does not update CHANGELOG.md, as that will be maintained by humans right now.

Co-authored-by: Radek Simko <radek.simko@gmail.com>
Co-authored-by: Daniel Banck <dbanck@users.noreply.github.com>