Copy patterns from HCL grammar to Terraform grammar (#17)
* Copy patterns from HCL grammar to Terraform grammar

This resolves an issue with nested include patterns. As soon as an
include hits the HCL grammar file, it would never include anything
from the Terraform grammar file. Because of this, highlighting of
functions would break in deeper nested blocks or function calls.

The commit introduces a lot of code duplication, but it all will go
away with the introduction of templating for grammar files.

* Update snapshot files for Terraform grammar