Fix highlighting of nested complex types (#20)
* add examples and snapshot tests (incorrect)
* Add list, set, map, tuple, and object as types to HCL

This also removes the incomplete definition for structural types,
which failed to highlight nested structures correctly. We now only
rely on highlighting through the type keywords.

* Add list, set, map, tuple, and object as types to Terraform

This also removes the incomplete definition for structural types,
which failed to highlight nested structures correctly. We now only
rely on highlighting through the type keywords.

Furthermore, we don't want to highlight list and map as functions,
so those keywords are removed from the functions list.