Build Syntaxes (#22)
* Syntax builder

A Golang application that takes TextMate grammar rules defined in source files, combines them, and produces grammar files for use in text editors.

Currently produces  HCL and Terraform grammar files that are exact replicas of current shipped versions.

* Format existing TextMate Grammar files

To aid in comparison, format the existing grammar files in the manner the builder will write them out as. This produces a large git diff now, but enables a smaller one later when the builder is run.

* Run builder on existing grammar

This uses the hcl and terraform source files to produce similar grammar files as those shipped currently. The Terraform source file has minor differences to show that the builder was run and produces syntactically correct files. The existing set of passing tmGrammar snapshot tests show that there are no build errors.