Add test for block names with non-english word characters
This originally started out as an attempt to exclude digits from the begining of the name by using "[A-Za-z]" but I then noticed that "[a-zA-Z]" only includes literally a-zA-Z (english letters), whereas \w includes all word characters.

https://go.dev/play/p/Z-iysCB4BRU

We could probably still exclude digits, but it doesn't seem as trivial as I thought it would be.