Be more permissive in allowed characters in a block label
Previously a block label would only be allowed to contain the set of
characters in an Indentifier, however a block label string literal can
contain pretty much any character. This commit updates the base HCL grammar
to support both indentifiers and string literals in block labels and adds tests
to ensure this behaviour works as expected.